chore: restructure function locations
